Ediaz/cp 28 improve shipment api services

+++ Transenvios.Shipping.Api/Domains/ShipmentOrderService/Requests/ShipmentOrderRequest.cs
@@ -3,18 +3,12 @@
     public class ShipmentOrderRequest
     {
         public ShipmentOrderRouteRequest? Route { get; set; }
-
         public IList<ShipmentOrderItemRequest>? Items { get; set; }
-
         public ShipmentOrderPersonRequest? Sender { get; set; }
-
         public ShipmentOrderPersonRequest? Recipient { get; set; }
-
-        public long BasePrice { get; set; }
-
-        public long Taxes { get; set; }
-
-        public long Total { get; set; }
+        public decimal BasePrice { get; set; }
+        public decimal Taxes { get; set; }
+        public decimal Total { get; set; }
         public string? ApplicantEmail { get; set; }
     }
 }
